refactor: 상금의 총합 자료형 변경 상금의 총합이 int 범위를 넘어설 수 있으므로 저장하고 계산하는 과정의 자료형들을 int에서 long으로 변경하였습니다.

+++ src/main/java/lotto/utility/Utility.java
@@ -37,15 +37,15 @@ public class Utility {
         return numbers;
     }
 
-    public int getProfit(int[][] totalResult) {
-        int sum = 0;
+    public long getProfit(int[][] totalResult) {
+        long sum = 0;
         Prize prize = new Prize();
         int[][] prizes = prize.getPrize();
 
         for(int i = 3; i <= 6 ; i++) {
-            sum += totalResult[i][0]*prizes[i][0];
+            sum += (long)totalResult[i][0] * (long)prizes[i][0];
         }
-        sum += totalResult[5][1]*prizes[5][1];
+        sum += (long)totalResult[5][1] * (long)prizes[5][1];
 
         return sum;
     }
